add InpatientWard controller
